MAINT: interpolate/RBF: unify "generic" and pythran RBF implementations
- Convert the pythran routines to use the same 'vectorized' implementations as the Array API backend
- Use the same source for pythran and "generic"
- Install both a pythranized extension and its source (for alternative array backends)
