+++ scipy/interpolate/_rbfinterp_np.py
@@ -1,25 +1,12 @@
 import numpy as np
+from scipy import linalg
 from numpy.linalg import LinAlgError
 from scipy.linalg.lapack import dgesv  # type: ignore[attr-defined]
 from ._rbfinterp_common import _monomial_powers_impl
 
 from ._rbfinterp_pythran import (
-    _build_system as _pythran_build_system,
-    _build_evaluation_coefficients as _pythran_build_evaluation_coefficients,
-    _polynomial_matrix as _pythran_polynomial_matrix
+    _build_system, _build_evaluation_coefficients, polynomial_matrix
 )
-
-
-# trampolines for pythran-compiled functions to drop the `xp` argument
-def _build_evaluation_coefficients(
-    x, y, kernel, epsilon, powers, shift, scale, xp
-):
-    return _pythran_build_evaluation_coefficients(
-        x, y, kernel, epsilon, powers, shift, scale
-    )
-
-def polynomial_matrix(x, powers, xp):
-    return _pythran_polynomial_matrix(x, powers)
 
 
 def _monomial_powers(ndim, degree, xp):
@@ -28,10 +15,6 @@ def _monomial_powers(ndim, degree, xp):
     if len(out) == 0:
         out = out.reshape(0, ndim)
     return out
-
-
-def _build_system(y, d, smoothing, kernel, epsilon, powers, xp):
-    return _pythran_build_system(y, d, smoothing, kernel, epsilon, powers)
 
 
 def _build_and_solve_system(y, d, smoothing, kernel, epsilon, powers, xp):
@@ -84,6 +67,7 @@ def _build_and_solve_system(y, d, smoothing, kernel, epsilon, powers, xp):
         raise LinAlgError(msg)
 
     return shift, scale, coeffs
+
 
 def compute_interpolation(x, y, kernel, epsilon, powers, shift, scale, coeffs, xp):
     vec = _build_evaluation_coefficients(
