+++ scipy/interpolate/_rbfinterp_pythran_src.py
@@ -1,39 +1,41 @@
-import numpy as np
+""" Internal primitives for RBF interpolation.
 
+Routines here are of dual purpose:
+- AOT compiled by pythran (hence # pythran export comments)
+- used for the "generic", Array API agnostic backend with non-numpy array libraries,
+  See `_rbfinterp_xp.py`.
+"""
 
-def linear(r):
+def linear(r, xp):
     return -r
 
 
-def thin_plate_spline(r):
-    if r == 0:
-        return 0.0
-    else:
-        return r**2*np.log(r)
+def thin_plate_spline(r, xp):
+    return xp.where(r == 0, 0, r**2 * xp.log(r))
 
 
-def cubic(r):
+def cubic(r, xp):
     return r**3
 
 
-def quintic(r):
+def quintic(r, xp):
     return -r**5
 
 
-def multiquadric(r):
-    return -np.sqrt(r**2 + 1)
+def multiquadric(r, xp):
+    return -xp.sqrt(r**2 + 1)
 
 
-def inverse_multiquadric(r):
-    return 1/np.sqrt(r**2 + 1)
+def inverse_multiquadric(r, xp):
+    return 1.0 / xp.sqrt(r**2 + 1.0)
 
 
-def inverse_quadratic(r):
-    return 1/(r**2 + 1)
+def inverse_quadratic(r, xp):
+    return 1.0 / (r**2 + 1.0)
 
 
-def gaussian(r):
-    return np.exp(-r**2)
+def gaussian(r, xp):
+    return xp.exp(-r**2)
 
 
 NAME_TO_FUNC = {
@@ -48,48 +50,30 @@ NAME_TO_FUNC = {
    }
 
 
-def kernel_vector(x, y, kernel_func, out):
-    """Evaluate RBFs, with centers at `y`, at the point `x`."""
-    for i in range(y.shape[0]):
-        out[i] = kernel_func(np.linalg.norm(x - y[i]))
-
-
-def polynomial_vector(x, powers, out):
-    """Evaluate monomials, with exponents from `powers`, at the point `x`."""
-    for i in range(powers.shape[0]):
-        out[i] = np.prod(x**powers[i])
-
-
-def kernel_matrix(x, kernel_func, out):
-    """Evaluate RBFs, with centers at `x`, at `x`."""
-    for i in range(x.shape[0]):
-        for j in range(i+1):
-            out[i, j] = kernel_func(np.linalg.norm(x[i] - x[j]))
-            out[j, i] = out[i, j]
-
-
-def polynomial_matrix(x, powers, out):
+# pythran export polynomial_matrix(float[:, :], int64[:, :], numpy pkg)
+def polynomial_matrix(x, powers, xp):
     """Evaluate monomials, with exponents from `powers`, at `x`."""
-    for i in range(x.shape[0]):
-        for j in range(powers.shape[0]):
-            out[i, j] = np.prod(x[i]**powers[j])
+    return xp.prod(x[:, None, :] ** powers, axis=-1)
 
 
-# pythran export _kernel_matrix(float[:, :], str)
-def _kernel_matrix(x, kernel):
+# NB: exported only to use in the tests
+# pythran export kernel_matrix_at_centers(float[:, :], str, numpy pkg)
+def kernel_matrix_at_centers(x, kernel, xp):
     """Return RBFs, with centers at `x`, evaluated at `x`."""
-    out = np.empty((x.shape[0], x.shape[0]), dtype=float)
     kernel_func = NAME_TO_FUNC[kernel]
-    kernel_matrix(x, kernel_func, out)
-    return out
+    return kernel_matrix(x, kernel_func, xp)
 
 
-# pythran export _polynomial_matrix(float[:, :], int64[:, :])
-def _polynomial_matrix(x, powers):
-    """Return monomials, with exponents from `powers`, evaluated at `x`."""
-    out = np.empty((x.shape[0], powers.shape[0]), dtype=float)
-    polynomial_matrix(x, powers, out)
-    return out
+def kernel_matrix(x, kernel_func, xp):
+    """Evaluate RBFs, with centers at `x`, at `x`."""
+    return _kernel_matrix_impl(x, x, kernel_func, xp)
+
+
+def _kernel_matrix_impl(x, y, kernel_func, xp):
+    """Evaluate RBFs, with centerx at `y`, at `x`."""
+    return kernel_func(
+        xp.linalg.vector_norm(x[None, :, :] - y[:, None, :], axis=-1), xp
+    )
 
 
 # pythran export _build_system(float[:, :],
@@ -97,8 +81,9 @@ def _polynomial_matrix(x, powers):
 #                              float[:],
 #                              str,
 #                              float,
-#                              int64[:, :])
-def _build_system(y, d, smoothing, kernel, epsilon, powers):
+#                              int64[:, :],
+#                              numpy pkg)
+def _build_system(y, d, smoothing, kernel, epsilon, powers, xp):
     """Build the system used to solve for the RBF interpolant coefficients.
 
     Parameters
@@ -128,38 +113,34 @@ def _build_system(y, d, smoothing, kernel, epsilon, powers):
         Domain scaling used to create the polynomial matrix.
 
     """
-    p = d.shape[0]
     s = d.shape[1]
     r = powers.shape[0]
     kernel_func = NAME_TO_FUNC[kernel]
 
     # Shift and scale the polynomial domain to be between -1 and 1
-    mins = np.min(y, axis=0)
-    maxs = np.max(y, axis=0)
+    mins = xp.min(y, axis=0)
+    maxs = xp.max(y, axis=0)
     shift = (maxs + mins)/2
     scale = (maxs - mins)/2
     # The scale may be zero if there is a single point or all the points have
     # the same value for some dimension. Avoid division by zero by replacing
     # zeros with ones.
-    scale[scale == 0.0] = 1.0
+    scale = xp.where(scale == 0.0, 1.0, scale)
 
     yeps = y*epsilon
     yhat = (y - shift)/scale
 
-    # Transpose to make the array fortran contiguous. This is required for
-    # dgesv to not make a copy of lhs.
-    lhs = np.empty((p + r, p + r), dtype=float).T
-    kernel_matrix(yeps, kernel_func, lhs[:p, :p])
-    polynomial_matrix(yhat, powers, lhs[:p, p:])
-    lhs[p:, :p] = lhs[:p, p:].T
-    lhs[p:, p:] = 0.0
-    for i in range(p):
-        lhs[i, i] += smoothing[i]
+    out_kernels  = kernel_matrix(yeps, kernel_func, xp)
+    out_poly = polynomial_matrix(yhat, powers, xp)
 
-    # Transpose to make the array fortran contiguous.
-    rhs = np.empty((s, p + r), dtype=float).T
-    rhs[:p] = d
-    rhs[p:] = 0.0
+    lhs = xp.concat(
+        [
+            xp.concat((out_kernels, out_poly), axis=1),
+            xp.concat((out_poly.T, xp.zeros((r, r))), axis=1)
+        ]
+    , axis=0) + xp.diag(xp.concat([smoothing, xp.zeros(r)]))
+
+    rhs = xp.concat([d, xp.zeros((r, s))], axis=0)
 
     return lhs, rhs, shift, scale
 
@@ -170,9 +151,10 @@ def _build_system(y, d, smoothing, kernel, epsilon, powers):
 #                          float,
 #                          int64[:, :],
 #                          float[:],
-#                          float[:])
+#                          float[:],
+#                          numpy pkg)
 def _build_evaluation_coefficients(x, y, kernel, epsilon, powers,
-                                   shift, scale):
+                                   shift, scale, xp):
     """Construct the coefficients needed to evaluate
     the RBF.
 
@@ -198,19 +180,18 @@ def _build_evaluation_coefficients(x, y, kernel, epsilon, powers,
     (Q, P + R) float ndarray
 
     """
-    q = x.shape[0]
-    p = y.shape[0]
-    r = powers.shape[0]
     kernel_func = NAME_TO_FUNC[kernel]
 
     yeps = y*epsilon
     xeps = x*epsilon
     xhat = (x - shift)/scale
 
-    vec = np.empty((q, p + r), dtype=float)
-    for i in range(q):
-        kernel_vector(xeps[i], yeps, kernel_func, vec[i, :p])
-        polynomial_vector(xhat[i], powers, vec[i, p:])
+    vec = xp.concat(
+        [
+            _kernel_matrix_impl(yeps, xeps, kernel_func, xp),
+            xp.prod(xhat[:, None, :] ** powers, axis=-1)
+        ], axis=-1
+    )
 
     return vec
